import datamodule to ensure registration
