+++ mttl/datamodule/__init__.py
@@ -1,55 +1,23 @@
-from collections import defaultdict
-
-import numpy as np
-from torch.utils.data.dataset import ConcatDataset
-
-
-def take_n_examples_per_task(task_names, n, rng=None):
-    """Returns indices of x / n examples per task given a list of task names.
-
-    Args:
-        task_names (list): List of task names, one per example.
-        n (int): Subsampling factor.
-        rng (np.random.RandomState, optional): Random number generator. Defaults to None.
-
-    Returns:
-        list: List of indices.
-    """
-    if rng is None:
-        rng = np.random.RandomState(0)
-
-    tasks_to_ids = defaultdict(list)
-    for i, task in enumerate(task_names):
-        tasks_to_ids[task].append(i)
-    indices = []
-    for task in tasks_to_ids.keys():
-        indices += rng.choice(
-            tasks_to_ids[task], max(len(tasks_to_ids[task]) // n, 1), replace=False
-        ).tolist()
-    return indices
-
-
-class TrainIndices:
-    _instance = None
-
-    def __init__(self, dataset, num_examples, seed):
-        self.num_examples = num_examples
-        self.seed = seed
-        self.rng = np.random.RandomState(self.seed)
-        self.train_indices = self.rng.randint(
-            0, len(dataset), self.num_examples
-        ).tolist()
-
-    @classmethod
-    def get(cls, dataset, num_examples, seed):
-        if cls._instance is None:
-            cls._instance = cls(dataset, num_examples, seed)
-
-        return cls._instance.train_indices
-
-
-class IndexConcatDataset(ConcatDataset):
-    def __getitem__(self, idx):
-        example_info = super().__getitem__(idx)
-        example_info["example_id"] = idx
-        return example_info
+from mttl.datamodule.arc_data_module import ArcDataConfig, ArcMultiChoiceDataModule
+from mttl.datamodule.codex_data_module import CodexDataConfig, CodexDataModule
+from mttl.datamodule.hellaswag_data_module import (
+    HellaswagDataConfig,
+    HellaswagMultiChoiceDataModule,
+)
+from mttl.datamodule.mmlu_data_module import MMLUDataConfig, MMLUDataModule
+from mttl.datamodule.mt_seq_to_seq_module import (
+    FlanConfig,
+    FlanModule,
+    FlatMultiTaskConfig,
+    FlatMultiTaskModule,
+)
+from mttl.datamodule.openbookqa_data_module import (
+    OpenbookQADataConfig,
+    OpenbookQAMultiChoiceDataModule,
+)
+from mttl.datamodule.piqa_data_module import PiqaDataConfig, PiqaMultiChoiceDataModule
+from mttl.datamodule.superglue_data_module import BoolQDataModule, SuperGLUEDataConfig
+from mttl.datamodule.winogrande_data_module import (
+    WinograndeDataConfig,
+    WinograndeMultiChoiceDataModule,
+)
